@@ -1,18 +1,28 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk.Web">
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
         <UserSecretsId>8fd7d6a4-9bfc-42ac-840b-dc145be7f1cc</UserSecretsId>
         <LangVersion>latest</LangVersion>
+        <Nullable>enable</Nullable>
+        <ProduceReferenceAssembly>true</ProduceReferenceAssembly>
     </PropertyGroup>
     <ItemGroup>
         <Folder Include="wwwroot\" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="NPoco" Version="3.9.4" />
-        <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="1.1.0" />
-        <PackageReference Include="AsyncAwaitBestPractices" Version="2.1.0" />
+        <PackageReference Include="AsyncAwaitBestPractices" Version="4.1.0-pre1" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.0.1" />
+        <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="3.1.0" />
+    </ItemGroup>
+    <ItemGroup>
+      <Content Remove="local.settings.json" />
+    </ItemGroup>
+    <ItemGroup>
+      <None Include="local.settings.json">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        <CopyToPublishDirectory>Never</CopyToPublishDirectory>
+      </None>
     </ItemGroup>
     <Import Project="..\TextMood.Shared\TextMood.Shared.projitems" Label="Shared" Condition="Exists('..\TextMood.Shared\TextMood.Shared.projitems')" />
     <Import Project="..\TextMood.Backend.Shared\TextMood.Backend.Shared.projitems" Label="Shared" Condition="Exists('..\TextMood.Backend.Shared\TextMood.Backend.Shared.projitems')" />
