@@ -3,8 +3,8 @@
     public static class EmojiConstants
     {
         public const string SadFaceEmoji = "☹️";
-        public const string NeutralFaceEmoji = "\U0001F610";
-        public const string HappyFaceEmoji = "\U0001F603";
-        public const string BlankFaceEmoji = "\U0001F636";
+        public const string NeutralFaceEmoji = "😐";
+        public const string HappyFaceEmoji = "😃";
+        public const string BlankFaceEmoji = "😶";
     }
 }
