@@ -4,11 +4,11 @@ namespace TextMood.Shared
 {
     public interface ITextMoodModel
     {
-		string Id { get; set; }
-		string Text { get; set; }
-        double? SentimentScore { get; set; }
-		DateTimeOffset CreatedAt { get; set; }
-		DateTimeOffset UpdatedAt { get; set; }
-		bool IsDeleted { get; set; }
+		string Id { get; }
+		string Text { get; }
+        double? SentimentScore { get; }
+		DateTimeOffset CreatedAt { get; }
+		DateTimeOffset UpdatedAt { get; }
+		bool IsDeleted { get; }
     }
 }
