@@ -2,21 +2,13 @@
 {
     public static class EmojiServices
     {
-        public static string GetEmoji(double? sentimentScore)
+        public static string GetEmoji(double? sentimentScore) => sentimentScore switch
         {
-            switch (sentimentScore)
-            {
-                case double number when (number >= 0 && number < 0.4):
-                    return EmojiConstants.SadFaceEmoji;
-                case double number when (number >= 0.4 && number <= 0.6):
-                    return EmojiConstants.NeutralFaceEmoji;
-                case double number when (number > 0.6):
-                    return EmojiConstants.HappyFaceEmoji;
-                case null:
-                    return EmojiConstants.BlankFaceEmoji;
-                default:
-                    return string.Empty;
-            }
-        }
+            double number when number >= 0 && number < 0.4 => EmojiConstants.SadFaceEmoji,
+            double number when number >= 0.4 && number <= 0.6 => EmojiConstants.NeutralFaceEmoji,
+            double number when number > 0.6 => EmojiConstants.HappyFaceEmoji,
+            null => EmojiConstants.BlankFaceEmoji,
+            _ => string.Empty,
+        };
     }
 }
