@@ -7,16 +7,10 @@ namespace TextMood
 {
     abstract class TextResultsService : BaseApiService
     {
-        #region Constant Fields
         readonly static Lazy<ITextModelApi> _textModelApiClientHolder = new Lazy<ITextModelApi>(() => RestService.For<ITextModelApi>(BackendConstants.GetEmotionResultsAPIUrl));
-        #endregion
 
-        #region Properties
         static ITextModelApi TextModelApiClient => _textModelApiClientHolder.Value;
-        #endregion
 
-        #region Methods
-        public static Task<List<TextMoodModel>> GetTextModels() => ExecutePollyHttpFunction(TextModelApiClient.GetTextModels);
-        #endregion
+        public static Task<List<TextMoodModel>> GetTextModels() => AttemptAndRetry(TextModelApiClient.GetTextModels);
     }
 }
